Add packages, begin to document game code, make run script
I added lovebone and tove to handle skeletons and rendering svg files (sprites from the original flash game).
I began documenting the game's scripts scene-by-scene to make things easier to locate and recreate/map in the future.
There's a lot.
I'll probably ending up using some code as references while changing the formatting of other stuff so that it works (I don't want to rewrite all the string definitions and stuff)

At this point it loads a blank black screen and prints "It didn't crash!" to the console.