Add config, more documentation, remove lovebone/tove
I used the default config file from the love website and changed a few values. It's weird how a 500x500 window in love isn't square (?).
Not sure what that's about.

Added an icon for the window
Documented the remaining functions which are attached to scenes but not buttons.

Restructured code analysis markdown object slightly.

Note: I'm going to have to do a whole lot of repackaging of game sprites/animations and make them into love2d animations / shaders and so on.