Main.lua -- Move shit around --  define some globals
I just can't make up my mind about where the folders should live :P

This push is in the midst of getting the main loop to do what I want, which (at this point) is to draw sprites as actors with several pieces which can each have an offset from the coordinates of the main actor.
For example, a dude can be at 0,0 and that's where his torso is, but his foot is at 5,5 from center, and it'll draw both of those.

I want to keep the main loop as slim as possible and store relevant game data in a single directory such that you could swap out the whole game with another in the same format. That's what the "data_structure" file is for, so the loader knows where to look for different things within the game folder.
Is not done yet but I have the idea I guess.

I want to make it easy to port flash games to lua, in this case one which has a lot of animations which use little pieces of armor tied to a character.

Turns out my dumb ass was re-running the game loop at the end. love.run is called automatically, I shouldn't have called it again.