Animations will be fun.....
I did my best to dump the animation data into lua.
The x and y coordinate system is WAY off, but I can still use the skews for rotation of things in each frame and just put in the locations myself.
Need to clean up z_index stuff.
I'll probably end up cleaning up the animation data by hand such that it starts with a base position and the next frames are just movement, not exact positions but a translation from where it was before.

I'm curious to see what it will look like moving around so maybe I'll do that first idk.