Moar animation stuff - Model looks ok.
It turns out the position values are actually good somewhat, I just need to divide by 2 for some reason (idk). Some positions and sizes are slightly off but this should still help a ton.
I'll be converting the animations to something usable... manually.
Oh dear.

At least I get part of the data to slap in...