Thinking....
I think I'll come back to the loading screen at some point and make it look good but for now move on to the main menu screen and dealing with things like being able to click on buttons, and those buttons behaving differently when hovered and so on.

I think I could handle that by adding an updater function which loops over a list of sprites and using their coordinates and sprite size checks if the mouse is within those coordinates (yay math)
Idk. and click stuff of course.
mmmmmmm
I wonder how the other games handle it?