Moved some stuff around.
In an effort to minimize clutter I've removed temporary files related to the abilities and temporary refactored code and moved them to another area on my hard drive. What they used to be will be remembered in commits before this one in case something happens to it on my hard drive.

Also consolidated the loading scene because I want the engine load to later be separated from the game load.

Did some work with the gpt model to swap out my PNG for the bars with something programmatic so I can keep this screen a single file.

Will seperate the engine loading from the game asset loading at some point so sonny will have its own cool loading screen and the engine can have a basic fast as fuck one.
Might have to split up the loading stages in the game's own loading function so that it can grab the assets for the loading screen itself separately and then use them right away to display a pretty loading screen.
Ahh, multi-stage loading.